backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
databases
aiomysql
python-dotenv
pydantic